+++ src/app/dashboard/page.tsx
@@ -1,4 +1,4 @@
-import { DashboardLayout, PageLayout } from '@components/PageLayout'
+import { DashboardLayout, PageLayout } from '@/app/components/layouts'
 import { DashboardNav } from '../components/navigations'
 
 // MAKE A INIT COMPONENT
